+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 11.3
 - Fix list sort, search, page, save when multiple lists on same page in ListOutput.fnc.php
 - Remove null values from URL in PreparePHP_SELF.fnc.php
 - We are inside a sentence, convert nouns to lowercase (except for German) in ListOutput.fnc.php & Prompts.php
+- No "Default for Teacher" checkbox if Type is "Office Only" in AttendanceCodes.php
 
 Changes in 11.2.3
 -----------------